requirements.txt
pymupdf>=1.24.0
pikepdf>=9.0.0
Pillow>=10.0.0
tkinterdnd2>=0.3.0
